extract types out to higher level
